+++ thermostat/otaupdates.h
@@ -2,6 +2,6 @@
 #define _CONFIG_H_OTAUPDATES_
 
 void IpAddressToStr(IPAddress * addr, char * res, int len);
-char checkOTA(char * addr, int port);
+char checkOTA(char * addr, uint16_t port);
 
 #endif
